\name{metaIbasam-package}
\alias{metaIbasam-package}
\alias{metaIbasam}
\docType{package}
\title{
metaIbasam
}
\description{
Metapopulation version of Ibasam
}
\details{
\tabular{ll}{
Package: \tab metaIbasam\cr
Type: \tab Package\cr
Version: \tab 0.0.2\cr
Date: \tab 2018-02-09\cr
License: \tab GPL2\cr
}
}
\author{
Cyril Piou <cyril.piou@cirad.fr>
Mathieu Buoro <mathieu.buoro@inra.fr>
}
\references{
}

\keyword{ package }
\seealso{
}
\examples{
}
