% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiomeMQC.R
\name{microbiomeMQC}
\alias{microbiomeMQC}
\title{Calculate 4 key reporting measures}
\usage{
microbiomeMQC(input_file_path, taxonomic_level, output_file_path)
}
\arguments{
\item{input_file_path}{Path to the .xlsx or .csv file.e.g. "C:\\Users\\joeblogs\\Desktop\\data.xlsx"}

\item{taxonomic_level}{The taxonomic level ("strain", "species", or "genus") that you are using.}

\item{output_file_path}{Path to save the output CSV file (file path must end with the file name e.g. "C:\\Users\\joeblogs\\Desktop\\MQC.csv"}
}
\description{
Calculate the 4 key reporting measures for the WHO International Reference Reagents for the microbiome
}
\details{
The 4 key reporting measures (using the example of species) are:

Sensitivity: How many species from the reagent are correctly identified.

Diversity: The total number of of species detected.

FPRA: The relative abundance of false positives.

Similarity: The Bray-Curtis dissimilarity in composition between the species profile reported and the 'ground truth' profile of the reagent.

For test data do: data(MQC_testdata)
}
